fix: iOS issue with status bar
It seems that, at least on iOS 13 and iPhone X and +, we have an issue
with the status bar and this plugin.

Currently, the plugin hides the status bar during the photo, and enable it
after. But the bar was not well placed and the display was broken.

By not hidding the bar we resolve this issue. It seems that the bar is hidden
byt the OS and the view for taking the photo. So no need to hide it of our own.

Tested on iPhone X iOS 13 and iPhone 7 iOS 12